Move completion related tests to their dedicated test folder
