+++ src/test/suite/completion/assistant-completion.test.ts
@@ -2,9 +2,9 @@ import * as assert from "assert";
 import * as vscode from "vscode";
 import * as sinon from "sinon";
 
-import * as getRecipesApiCall from "../../graphql-api/get-recipes-for-client";
-import * as usedRecipeApiCall from "../../graphql-api/use-recipe";
-import { Language } from "../../graphql-api/types";
+import * as getRecipesApiCall from "../../../graphql-api/get-recipes-for-client";
+import * as usedRecipeApiCall from "../../../graphql-api/use-recipe";
+import { Language } from "../../../graphql-api/types";
 import {
   wait,
   closeFile,
@@ -26,7 +26,7 @@ import {
   javaRecipe,
   documentJavaRecipeImportsAfterPackageExpected,
   documentJavaRecipeImportsBetweenCommentsExpected,
-} from "../testUtils";
+} from "../../testUtils";
 
 // test recipe auto complete capabilities of the plugin, we create mocks and stub
 // for recipe fetch and recipe usage endpoints
