+++ src/test/suite/completion/variable-transformation.test.ts
@@ -2,9 +2,9 @@ import * as assert from "assert";
 import * as vscode from "vscode";
 import * as sinon from "sinon";
 
-import * as getRecipesApiCall from "../../graphql-api/get-recipes-for-client";
-import * as usedRecipeApiCAll from "../../graphql-api/use-recipe";
-import { Language } from "../../graphql-api/types";
+import * as getRecipesApiCall from "../../../graphql-api/get-recipes-for-client";
+import * as usedRecipeApiCAll from "../../../graphql-api/use-recipe";
+import { Language } from "../../../graphql-api/types";
 import {
   wait,
   closeFile,
@@ -14,7 +14,7 @@ import {
   recipeWithTransformVariables,
   testDataUri,
   updateConfig,
-} from "../testUtils";
+} from "../../testUtils";
 
 // test there's no recipe variable in the final recipe insertion, we create mocks and stub
 // for recipe fetch and recipe usage endpoints
